Moved rocket stuff into its own module
