+++ src/service/mod.rs
@@ -1,0 +1,3 @@
+mod rocket;
+
+pub use self::rocket::*;
