+++ src/service/rocket/mod.rs
@@ -1,0 +1,13 @@
+mod api;
+mod serve;
+
+pub mod server;
+
+#[cfg(test)]
+mod api_tests;
+#[cfg(test)]
+mod swagger;
+#[cfg(test)]
+mod test;
+#[cfg(test)]
+mod web;
