+++ src/service/rocket/swagger.rs
@@ -1,6 +1,7 @@
+use super::test::get_test_environment;
+
 #[test]
 fn test_index() {
-	use crate::test::get_test_environment;
 	use rocket::http::Status;
 	let env = get_test_environment("swagger_index.sqlite");
 	let client = &env.client;
@@ -10,7 +11,6 @@ fn test_index() {
 
 #[test]
 fn test_index_with_trailing_slash() {
-	use crate::test::get_test_environment;
 	use rocket::http::Status;
 	let env = get_test_environment("swagger_index_with_trailing_slash.sqlite");
 	let client = &env.client;
