+++ src/service/rocket/web.rs
@@ -1,8 +1,8 @@
+use super::test::get_test_environment;
+use rocket::http::Status;
+
 #[test]
 fn test_index() {
-	use crate::test::get_test_environment;
-	use rocket::http::Status;
-
 	let env = get_test_environment("web_index.sqlite");
 	let client = &env.client;
 	let response = client.get("/").dispatch();
